Cleanup error handling
- Stop depending on my unofficial for for graceful shutdown.
- Use thiserror and http-api-problem to report better errors.
