+++ src/start/trace.rs
@@ -1,0 +1,24 @@
+use axum::http::Request;
+use tower_http::request_id::{MakeRequestId, RequestId};
+use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
+use uuid::Uuid;
+
+#[derive(Clone, Copy)]
+pub(crate) struct RequestUuidService;
+
+impl MakeRequestId for RequestUuidService {
+    fn make_request_id<B>(&mut self, _request: &Request<B>) -> Option<RequestId> {
+        let request_id = Uuid::new_v4().to_string().parse().unwrap();
+        Some(RequestId::new(request_id))
+    }
+}
+
+pub(crate) fn init_tracing() {
+    tracing_subscriber::registry()
+        .with(tracing_subscriber::EnvFilter::new(
+            std::env::var("RUST_LOG")
+                .unwrap_or_else(|_| "cargo_lambda=info,tower_http=info".into()),
+        ))
+        .with(tracing_subscriber::fmt::layer())
+        .init();
+}
